Shift tab parser method to be part of the util file
